\name{makeappendix}
\alias{\preformatted{\makeappendix}}
\title{
Marks the beginning of an appendix
}
\description{
Places appendix on its own page and writes the user provided \bold{APPENDIX NAME} at the top of the page. Can be used to add multiple appendix sections.
}
\usage{
\preformatted{\makeappendix{Appendix Name}}
TEXT
}

