\name{maketable}
\title{
Marks the beginning of a table and its title
}
\description{
Used immediately before R code block which generates the table. Sets the alignment and font-size within the table. Following the code block, the table should end with: \preformatted{\stoptable}
}
\usage{
\preformatted{\maketable{Table Title}}
\bold{R Code Block}
\preformatted{\stoptable}
}

