\name{startbib}
\title{
Marks the beginning of the bibliography section
}
\description{
Places bibliography on its own page and writes "References" at the top of the page. Formats sources according to APA hanging indentation style. Usage \startbib
}
\usage{
\preformatted{\startbib}
Citation 1
Citation 2
Citation 3
    .
    .
    .
}

